backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.30
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
ccxt>=4.3.0
anthropic>=0.71.0
openai>=2.6.0
pydantic>=2.10.0
python-dotenv==1.0.0
psycopg[binary]>=3.2.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
greenlet>=3.0.0
websockets==12.0
numpy>=1.24.0
pandas>=2.0.0
TA-Lib>=0.6.0
email-validator>=2.0.0
deepmem>=0.1.0